Place every XCI file in a separate directory
fusesoc provides an option to run a synthesis from the source directory
(by passing the --no-export command line). This mode of operation is
great during development, as files can be edited and Vivado picks up the
changes immediately without another fusesoc run.

For this to work all XCI files must be in a separate directory. This is
a Vivado requirement, which reports the IP cores as "locked" if more
than one XCI file is found in the same directory.

This commit places each XCI file in its own directory, and adds
.gitignore files to ignore the output files Vivado writes into the
source tree.